move ast into a separate folder
Caleb don't murder me for this, if I had to keep scrolling through a 300+ line file to find what I wanted every 5 seconds I would go insane. Feel free to pr a "fix" after I don't have to do that anymore.
